Modify Py_buffer for CHERI support, breaking ABI3.
The Py_buffer structure is part of the Python ABI3, however the
The size of the structure changes with CHERI and the
alignment of the fields is not enough to prevent padding.
Make the padding explicit and reorder fields to make it less awkward
to initialize.
Making padding explicit should allow interoperability between the hybrid
and purecap python.